+++ CHANGELOG.rst
@@ -2,6 +2,8 @@ Changelog
 =========
 
 * 2.1 (not released)
+    * Django 1.7 and 1.8 support
+    * Add South and Django migrations
     * Changed Gravatar link to use HTTPS by default
     * Fixed a bug where the admin avatar list page would only show a user's primary avatar
     * Updated render_primary view to accept usernames with @ signs in them